refactor: add comment dialog 분리
